Fix match definitions of network policies in docs and CD
This change fixes an error in the match definitions of network policies.

Before the fix, the Kubernetes would parse the destination port and source
address as two different rules, instead of combining them to one.

This would accidentally cause the port to be accessible from any source
address (from a network policy perspective) and the defined source address
being able to access any ports on the matched pods.