Added information about pv type hostPath
Added a recommendation to not run a PV of type hostPath if they want to actually use an external volume that can be used by different nodes. Comes from ticket 11657 where the customer's pod was evicted to a different node during a planned maintenance window and thought their data was lost.